==> 01_data_understanding_user.ipynb <==
import os
import pandas as pd 

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', None)
pd.set_option('display.max_colwidth', None)

df_ratings = pd.read_csv(os.path.join('data', 'prepared_ratings.csv'))
df_ratings.info()

df_ratings['date'] = pd.to_datetime(df_ratings['date'])

user_analysis = df_ratings.groupby('userId').agg(
    unique_movies=('id', 'nunique'),  # Count of unique movie IDs
    mean_rating=('rating', 'mean'),  # Mean rating
    max_rating=('rating', 'max'),    # Maximum rating
    min_rating=('rating', 'min'),    # Minimum rating
    max_date=('date', 'max'),   # Most recent rating date
    min_date=('date', 'min')    # Earliest rating date
).reset_index()

user_analysis.head()

# Sort user_analysis by unique_movies in descending order
user_analysis = user_analysis.sort_values(by='unique_movies', ascending=False)
user_analysis.head(5)

user_analysis_top10 = user_analysis.head(20)
print("Shape of top 10 users analysis:", user_analysis_top10.shape)
print("Shape of original user analysis:", user_analysis.shape)
df_movies_top10 = df_ratings[df_ratings['userId'].isin(user_analysis_top10['userId'])]
print("Shape of movies rated by top 10 users:", df_movies_top10.shape)
print("Shape of original movies dataframe:", df_ratings.shape)

# Filter df_movies_top10 for ratings >= 4
df_movies_top10_high_ratings = df_movies_top10[df_movies_top10['rating'] >= 4]
print("Shape of filtered DataFrame:", df_movies_top10_high_ratings.shape)

df_movies_top10_high_ratings['date'] = pd.to_datetime(df_movies_top10_high_ratings['date'])

user_analysis_top10_high_ratings = df_movies_top10_high_ratings.groupby('userId').agg(
    unique_movies=('id', 'nunique'),  # Count of unique movie IDs
    mean_rating=('rating', 'mean'),  # Mean rating
    max_rating=('rating', 'max'),    # Maximum rating
    min_rating=('rating', 'min'),    # Minimum rating
    max_date=('date', 'max'),   # Most recent rating date
    min_date=('date', 'min')    # Earliest rating date
).reset_index()
user_analysis_top10_high_ratings.head(10)
